projet_video/requirements.txt
flask 
moviepy 
mysql-connector-python 
python-dotenv 
imageio 
imageio-ffmpeg 
numpy 
pillow 
